This latest version of new text adventure engine includes a properly finished "read" verb, the beginnings of a "combine --  with --" verb, new functions that allow the proper handling of weight when an object moves in or out of inventory (solving the problem when you want to pick up and drop things with no fuss as part of special behavior), a new function to strongarm what the current amount carried is, in a pinch, and a drop all/drop everything option. I also created shortcuts and set things up for quick navigation to any part of the program.
Addendum: Zelia playtests. As a result, created the synonym for examine, look, fixed what happens when there's no verb and no direct object. Fixed the incorrect in-game description of the balcony as being on the east. Added "go south" and "south" as alternatives to "s". Added "show verbs", "verbs", and "list verbs". Looking pretty solid.